requirements.txt
# Trabaholink - Tested Working Requirements
# Simplified and verified to build successfully

# Core Django
Django==5.1.6
djangorestframework==3.15.2
django-allauth==65.4.1
django-cors-headers==4.7.0

# ASGI & WebSocket
asgiref==3.8.1
daphne==4.1.2
channels==4.2.0
channels_redis==4.2.1

# Async & Celery
celery==5.3.4
redis==5.2.1

# Database
psycopg2-binary==2.9.10

# Geospatial (install after GDAL system packages)
GDAL==3.6.2
geopy==2.4.1

# Image Processing
numpy==1.24.4
opencv-python-headless==4.10.0.84
Pillow==11.1.0

# OCR - Basic (without torch/EasyOCR to avoid conflicts)
pytesseract==0.3.13

# Face Recognition - Temporarily disabled (add on server with fast internet)
# dlib==19.24.2
# face-recognition==1.3.0

# QR Code
pyzbar==0.1.9

# Browser Automation
playwright==1.49.1

# Email
django-anymail==12.0
sendgrid==6.11.0

# Environment
python-dotenv==1.0.1

# Utilities
requests==2.32.3
certifi==2025.1.31

# Twisted
Twisted==24.11.0
autobahn==24.4.2
